better finding of last page
but there is a bug here! it will take the page where the next topic is.

this works for 225 because the next topic could be half way down the page

but for 220 the next topics always start on a new page
so for 220 the content will always include some of the next chapters 😢